+++ stdlib/SparseArrays/src/SparseArrays.jl
@@ -1,7 +1,5 @@
 # This file is a part of Julia. License is MIT: https://julialang.org/license
 
-__precompile__(true)
-
 """
 Support for sparse arrays. Provides `AbstractSparseArray` and subtypes.
 """